utils/counting/inference/dip-vision-ds-ds-dev-refactor/requirements.txt
numpy>=1.26.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
python-dotenv==1.0.1
PyYAML==6.0.2
supervision>=0.17.1
ultralytics==8.0.158
